MB-25785: Ask a DCP client with diverging branch @ seqno 0 to rollback
It is possible that a DCP client can have a diverging branch (vb_uuid
mismatch) at seqno 0. Even though, it is not terribly incorrect to
go on with the DCP stream request without asking the client to rollback
when start_seqno is 0, recently a backup client has faced an issue with
this.

The commit fixes the bug by asking the client to rollback even though
it has a diverging branch at seqno 0. Also test cases are added for
the same and some of the exisiting test cases which didn't care about
the vb_uuid when stat_seqno is 0 are updated accordingly.

Also the commit ensures that, with this change, we do not end sending
rollback to clients with matching vb_uuid and start_seqno == 0 but
start_seqno < purge_seqno. That is, the behavior here is retained
as before.

Change-Id: I19759ab3ea5f9e2995b335b791fd024f63bb7242
Reviewed-on: http://review.couchbase.org/85814
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Mike Wiederhold <mike@couchbase.com>
Tested-by: Mike Wiederhold <mike@couchbase.com>