MB-26828: Fix race condition when operating on the same document
In order to allow atomic operations subdoc needs to use ADD in
order to add the document to the underlying engine (for SET
operations as a set operation with CAS == 0 would overwrite
the existing item created from a racing thread).

subdoc_fetch should have detected this and mapped the "not
stored" response from the underlying engine to EEXISTS in
a pure "SET" context.

Change-Id: Ia42b1b0ad7152a73873e00a10916670284344fbf
Reviewed-on: http://review.couchbase.org/85839
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>