testapp: Move test_subdoc*() functions to SubdocTestappTest
Move the various test_subdoc free functions to be member functions of
SubdocTestappTest. This is in preparation to be able to change the
behaviour of these functions based on the current test config (e.g. if
Hello::JSON is enabled/disabled) without having to pass an additional
argument down various call chains.

Change-Id: I0a8b034735001d3f8b44a3d3326af860bf0cfdbb
Reviewed-on: http://review.couchbase.org/86836
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>