MB-25785: Ask a DCP client with diverging branch @ seqno 0 to rollback
It is possible that a DCP client can have a diverging branch (vb_uuid
mismatch) at seqno 0. Even though, it is not terribly incorrect to
go on with the DCP stream request without asking the client to rollback
when start_seqno is 0, recently a backup client has faced an issue with
this.

The commit fixes the bug by asking the client to rollback even though
it has a diverging branch at seqno 0. Also test cases are added for
the same and some of the exisiting test cases which didn't care about
the vb_uuid when stat_seqno is 0 are updated accordingly.

Also the commit ensures that, with this change, we do not end sending
rollback to clients with matching vb_uuid and start_seqno == 0 but
start_seqno < purge_seqno. That is, the behavior here is retained
as before.

Further, this commit ensures that the DCP client in the server, that
is the DCPConsumer, opens/reconnects a stream with 'vb_uuid == 0' when
the start_seqno of the stream is 0.

Change-Id: I83bae0104c8b4dfc7a9e4abd2479733beaf92096
Reviewed-on: http://review.couchbase.org/85917
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>