Tighten up engine API; Require cookie for get_stats
Use gsl::not_null for the cookie parameter passed into
the engine and the add_stats callback (as the core expects
it to be a valid cookie).

Change the two parameter key + keylen into a cb::const_char_buffer

A followup to this patch (and the other not_null patches) change
the void pointer to a Cookie object passed by reference.

Change-Id: I31c72e4665ee548a3bb525b28b3a56b663c8f2c0
Reviewed-on: http://review.couchbase.org/85965
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>