MB-26021: Optimize sorting of queued items
When sorting queued items in the flusher (before writing to disk), we
spend a non-trivial amount of time comparing Items by key & seqno.

This patch optimizes it by combining the less-than and equality check
into a single call to compare(). This essentially halves the amount
time the actual compare takes.

Add a benchmark to measure the performance of
CompareQueuedItemsBySeqnoAndKey to measure this. This shows a 25%
reduction in runtime for the sorting process:

Before:

    Run on (8 X 2300 MHz CPU s)
    2017-11-28 15:41:11
    --------------------------------------------------------------------------
    Benchmark                                   Time           CPU Iterations
    --------------------------------------------------------------------------
    BM_CompareQueuedItemsBySeqnoAndKey    4541594 ns    4540907 ns        140

After:

    Run on (8 X 2300 MHz CPU s)
    2017-11-28 15:48:54
    --------------------------------------------------------------------------
    Benchmark                                   Time           CPU Iterations
    --------------------------------------------------------------------------
    BM_CompareQueuedItemsBySeqnoAndKey    3368195 ns    3365729 ns        221

Change-Id: I428179c648b8bb311215b22def5520dcb5f7983a
Reviewed-on: http://review.couchbase.org/86123
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>