MB-22010: [2/N] Implement Tagged Pointer
To improve our hash table eviction algorithm we need to hold more state
regarding when and how often documents in the cache are accessed.

However we do not want to increase the memory usage, especially as the
increase will be multiplied by every entry held in the hash table.

Given we maintain a 64-bit pointer to a StoredValue for each entry in
the hash table, on x86-64 architectures we can use the top 16-bits for
our own use.  This is because x86-64 only use the bottom 48 bits for
addressing.

This patch introduces a TaggedPtr class, which has the ability
to act as a pointer, as well as hold 16-bit of user data.

Change-Id: I26e615f62ed4fd87b97772ac33ddba4053b4e63a
Reviewed-on: http://review.couchbase.org/86177
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>