MB-22010: [3/N] Do not directly refer to a RCValue
As part of the TaggedPtr work we want SingleTheadedRCPtr to be applied
to a TaggedPtr.  To achieve this we need to avoid the
SingleThreadedRCPtr using a static_cast to directly access a RCValue.

This patch introduces a getRCValue function that returns a reference to
a RCValue.  This allows all the static_cast<RCValue*> to be removed
from the RCPtr and SingleThreadedRCPtr classes.

Change-Id: I1327a8069b9fa0f7857af959126708c456a34162
Reviewed-on: http://review.couchbase.org/86182
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>