MB-26021: Add TransactionContext to KVStore [1/2]
When commiting a transaction in the KVStores, various state needs to
be maintained to ensure that ep-engine is called back when items have
successfully been persisted to disk. This is typically handled by the
store maintaining a container of PersistenceCallback objects to be
invoked when the transactionis commited. However much of the state of
PersistenceCallback is identical across all operations in a commit -
for example the VBucket, the Stats object to update.

To reduce the memory cost of maintaing this state, introduce a new
TransactionContext object. Callers of KVStore pass an instance of this
into begin(), and it is owned by the KVStore until commit() or
rollback() is called. Upon commit, all PersistenceCallacks are invoked
with a reference to the TransactionContext; allowing them to make use
of the state stored there.

This initial patch just adds the instructure for the
TransactionContext; it is not yet used by ep-engine.

Change-Id: I8c3704c968270a871849e9c6b5efda22b221d79b
Reviewed-on: http://review.couchbase.org/86285
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>