Refactor: Move the implementation of adjust_timeofday into the core
It used to be implemented as a separate shared object which we
would load at runtime during execution of unit tests. That piece
of code complicates the core, as we need to have support for
registering and perform mapping of commands. This generic
functionality was broken as part of the RBAC task, as we'd
fail the opcode unless it had an entry in the privilege
check.

By including the adjust_timeofday impl into the core we may
rip out the functionality for loading generic protocol extensions
(which we've never used in production anyway)

Change-Id: I0c37cf5eaca3401ff5baab83eeeb3bbc076c4575
Reviewed-on: http://review.couchbase.org/86305
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>