MB-27037: Remove unnecessary uses of item ptr in store()
Engine API store() method takes an optional 'outItem' parameter to
return the item stored. However, many instances of the code don't use
it; and simply immediatley release the item.

This is both unnecessarily complex, but also error-prone, as callers
need to check that the item is non-null before calling release (it
could be null if the store() failed).

Remove all instances of the outItem parameter where is it unused.

Change-Id: I149321f69bd3e8697c21cd2a62c0656589fbf091
Reviewed-on: http://review.couchbase.org/86311
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>