MB-27099: Don't try to dereference stl containers in connections stats
The connections stat don't try to lock the object it inspects,
so if the connection is used by a different worker thread the
containers could be resized or modified behind our back.
The code tried to access the elements, which could have been
invalidated and causing a crash.

Given that the details is of limited value, lets just drop
dumping them.

Change-Id: I411d3dceacdbf59541c03ce4004a9cedf6d8c4c5
Reviewed-on: http://review.couchbase.org/86359
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>