MB-27063: Add user filter to Audit Daemon
Increase to version 2 of the audit configuration, which has the addition
of a new attribute - disabled_users.  This is an array of strings,
where each string is a user.

If an event with a real_userid:user or effective_userid:user
matches with a user in the disabled_users, the event will not
be outputted to the logs.

Change-Id: Ia62c0e7c73f51f11c86416374b08d3286f11952b
Reviewed-on: http://review.couchbase.org/86197
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>