MB-26021 [1/6]: Limit #backfill items flushed in a single batch
Add a new configuration parameter - flusher_backfill_batch_limit -
which allows the flusher to contrain how many backfill items will be
flushed in a single batch. This defaults to zero, which means no limit
and hence behaves as previously.

If a non-zero value is specified then no more than that number of
backfill items will be added to a single vBucket flusher commit; the
given vBucket will be flagged to indicate there's more data available
and hence flusher should be re-scheduled.

Change-Id: Ic9c86f915f63fca7f8802cc40597907b5a0c1d2b
Reviewed-on: http://review.couchbase.org/89341
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>