MB-27105 [RocksDB] Set the baseline size of the 'local' Memtable to 4KB
The 'local' Column Family contains only a few data compared to the other
CFs, but on a long running cluster the 'local' Memtable will be filled
up to the maximum (currently all the CFs are set to the same maximum).
This may lead to a lot of memory allocated for the 'local' Memtable.
To prevent this, we set the baseline size the 'local' Memtable to 4KB.

Change-Id: I2fee1bce271bba9d08219afb90ae5755a99cc79b
Reviewed-on: http://review.couchbase.org/86383
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>