MB-22010: [5/N] Implement Tagged Pointer for SingleThreadRCPtr
This patch modifies the SingleThreadedRCPtr class, to allow the
TaggedPtr class to be provided as the pointer type.  This allows
the reference count properities of SingleThreadedRCPtr to be combined
with the tagging properties of TaggedPtr.

Also adds a static helper method to TaggedPtr for setting the tag of
a TaggedPtr that is wrapped inside a Pointer, such as
SingleThreadedRCPtr.

Change-Id: I01b5696c52523fec35626ea18bc34b3c3ba9df9a
Reviewed-on: http://review.couchbase.org/86324
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>