Put tracing flag in connection object
We'll be able to operate on multiple commands in parallel
and hence have multiple cookie objects for each connection.

The lifetime for the cookie object is from the command starts
executing until the command is done (currently we only reset
the object to avoid extra memory allocations which isn't really
needed).

This patch initialize the cookie's trace flag every time a
new command is started (once we start using multiple cookies
the flag in the cookie should be const and initialized through
the constructor)

Change-Id: I25745947c2356340e7805f281b3b6c10a615cc49
Reviewed-on: http://review.couchbase.org/86563
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Premkumar <premkumar.thangamani@couchbase.com>