testapp: Fix ep-engine-getset-unit-tests.full_eviction
memcached_testapp supports running with three engine types: default,
ep & ep_full_eviction (ep-engine with full-eviction enabled). This is
configured in McdEnvironment(), which sets EpBucketImpl::fullEviction
to true in the case of ep_full_eviction.

ep-engine-getset-unit-tests.full_eviction is one such test which
attempts to run in ep_full_eviction mode.

However, EpBucketImpl::fullEviction is effictively ignored in
setUpBucket() - the settings string it creates isn't used as part of
the bucket config. As such it has no effect.

Fix this.

Change-Id: I514066610e8c4492f6481ef687f8ba888ea88f29
Reviewed-on: http://review.couchbase.org/86578
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>