MB-27173: Command timings should not include copy to kernel
The command timings should not include the time it takes to
copy the potential response into the kernels send buffers.

The motivation for this change is that we're moving towards
using the bufferevents in libevent and we don't want to
allocate a separate bufferevent buffer to get the callback
from to stop the timer.

As part of the move over to bufferevent we'll also simplify
the state machinery so that we'll be able to start processing
the next command before the previous command is completely
sent to the client, but stop processing input commands when
the send buffer exceed a certain size.

Given that the timing represent a single command it makes
sense to simplify the code by keeping the timer inside the
command context rather than the connection object.

Change-Id: Ic2f82cb97be4ae63ea92d83ab3be54f2bfd44f82
Reviewed-on: http://review.couchbase.org/86612
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>