MB-27174: Remove releaseCookie call from dcpOpen and move reserve
1. Reserve the cookie only on success, reserve means that the engine
is storing a copy of the cookie (so we must inc the ref-count).

2. Remove the explicit call to release from dcpOpen - a frontend
worker thread must not call release-cookie.

These changes have had to be manually tested due to the limited
'full-stack' dcp testing available.

* Code was changed to force the if (handler == nullptr) error to
trigger - and the deadlock was observed (debug attach showed 1 thread
stuck). Removing the releaseCookie removed the deadlock.

Change-Id: If845a4c338f22842c3c10bfcce58fc00e52ad6e5
Reviewed-on: http://review.couchbase.org/86657
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>