MB-27123 [RocksDB] One DB per Shard, model VBuckets with CFs
The current RocksDBKVStore models every VBucket as a separated RocksDB
instance. But, some bits of the RocksDB Wiki suggest that too many
databases may hurt the performance, in particular when WriteBatch and
sync writes to WAL are issued, which essentially what we do.
Thus, with this change we move from one DB per VBucket to a unique DB
per Shard, and we model a VBucket as a set of three ColumnFamilies
(default, seqno and local CFs per VBucket).

Change-Id: I3ffc0de34bf479a9d1b7d6107841aba11244ca7f
Reviewed-on: http://review.couchbase.org/86710
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>