MB-2210: [6/N] Remove UniqueTaggedPtr wrapper class
UniqueTaggedPtr was introduced as a wrapper for when TaggedPtr is
wrapped in a std::unique_ptr.  However with the introduction of the
TaggedPtr helper method "updateTag", (which takes a pointer that wraps
a TaggedPtr and a value that is to be stored in the tag), the
UniqueTaggedPtr is now of limited value.  Also it duplicates the
interface of std::unique_ptr.

Therefore this patch removes the class and modifies the code that uses
it to directly operate on a std::unique_ptr.

Change-Id: I9617ce43a9cd33e8d385a9a00af97bab5d44c6bc
Reviewed-on: http://review.couchbase.org/86722
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>