MB-27215: Clean-up KVStore keyStats/kvstats
Found that the publically declared structure is only ever used by
private methods of couch-kvstore, further to that parts of the
kvstats_entry are written but never read, e.g. the second part
of the std::pair and the vbucket member of the KeyStats structure.

Note:
The internal std::unordered_map (now maps key to exists bool) could be
enhanced further to be a std::unordered_set storing only the existing
keys, a future change may need to map key to more than just bool.

Change-Id: I374ce15f5da3f8248cff68fa7dc45545054b6f53
Reviewed-on: http://review.couchbase.org/86770
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>