MB-27224 [RocksDB] Fix 'out_of_range' error on CF aggregate stats
The numeric RocksDB stats from Properties are of type uint64_t, so
we need to use 'std::stoull' for the string-to-numeric conversion.
Also, when we get ColumnFamily aggregate stats we need to initialize the
output value to 0 because it is passed as a non-initialized size_t& by
the caller (KVBucket::getKVStoreStat).

Change-Id: I8d76e815ed189323e0d5e05f1440d7045fd53ced
Reviewed-on: http://review.couchbase.org/86819
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>