MB-27223: Don't reset NRU on an update
There has been a regression from 4.6.2 where on an update we reset the
NRU.  This patch reverts the regression and adds a specific test for
when an update is performed.

Argurably the NRU should be modified to reflect that the value is in
the working set.  However this improvement is not part of MB-27223.

Change-Id: Ic3ba4ac0243336e00829ed7e48e8bc481a2c63f1
Reviewed-on: http://review.couchbase.org/86895
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>