MB-27063: Add config flag to enable / disable filtering
Add a global configuration option "filtering_enabled" which if set to
true allows audit events to be filtered by user (from real_userid or
effective_userid).

The option is is required to be defined in version 2 of the audit
configuration (similar to the diabled_users list).

Also adds an associated test.

Change-Id: Ibee723a41100e999c3fafdcea416722e395f20e5
Reviewed-on: http://review.couchbase.org/86972
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>