MB-26248: testapp: remove datatype=JSON from mutation requests
We have a number of instances where tests set request.datatype.JSON in
mutations. This is (a) only valid if the client has negotiated
Hello::JSON; and (b) even then redundant, as the server will ignore
the value of datatype.JSON and re-detect itself.

Currently the vast majority of tests run with Hello::JSON true;
however we want to shorlty change that to test clients with and
without Hello::JSON. As such, remove request.datatype=JSON from such
tests in preparation for this.

Change-Id: I865e145a68681bfa8cee9987ffd38fefc13bb536
Reviewed-on: http://review.couchbase.org/86902
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>