MB-26248: testapp: Make Hello::JSON optional
Currently most testapp tests always enable Hello::JSON. We want to
make this optional, so we can test the behaviour of the server is
correct both when it is on and when it is off.

This patch adds a virtual method to TestappTest - hasJSONSupport() -
which subclasses can override as appropriate if they want to enable
JSON or not. For example, parameterized tests could override that and
return a value based on GetParam().

The default value of hasJSONSupport() is No - this is to match the
current/legacy expected behaviour where JSON is not successfully
negotiated.

Change-Id: I87f4d96db900686b3b797441f222af0d8b79e3e6
Reviewed-on: http://review.couchbase.org/86906
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>