MB-27326 [RocksDB] Merge the Local CF into the Seqno CF
A recent analysis on Space Amplification showed that setting the same
size for the Default, Seqno and Local Memtables prevents old WAL files
from being deleted. This happens because the WAL is shared among CFs,
and the Seqno and Local CFs store far less data than the Default CF.
To address this issue, we essentially set the size of the Memtable for
the three CFs proportionally, i.e. minimum size for the Local Memtable
(MB-27105) and possibility to set the Default and Seqno ones from
configuration (MB-27175).
The analysis showed that the last step toward the minimum Space
Amplification would be flushing the Local Memtable periodically. But,
that would trigger compaction more often and increase write
amplification. Also, some Perf Tests under Universal Compaction showed
that compaction is not always triggered when
'level0_file_num_compaction_trigger' is reached
(https://github.com/facebook/rocksdb/wiki/Universal-Compaction). That
causes the accumulation of many tiny L0 files and a consequent
Stalling of Writes.
Thus, we address both of the issues above merging the Local CF into
the Seqno CF. Doing so we do not need to implement any periodic flush
of the LocalCF and we prevent Writes Stalling.

Change-Id: If6260feb1f67ea72ff85391e944b8d5f859dbb23
Reviewed-on: http://review.couchbase.org/87011
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>