Add possibility to copy request packet to Cookie
To simplify the logic for the out of order execution mode
we should allow the cookie to copy the content of the packet
into it's own buffers.

The motivation for this is that when using bufferevents the
input buffer may get reallocated when it tries to read more
data (or as part of a "pullup"). If the commands previous
in the pipe returned "ewouldblock" they _could_ contain
pointers which is no longer valid.

The intention is for the core to first try to run the command
with the data pointing into bufferenvents buffers, but if
the command blocks for some reason we need to create a
copy of the buffer so that we can pull up the next packet
from the bufferevent and try executing that.

Change-Id: I67c966303163ff09b952e203076fc8a2c9784a77
Reviewed-on: http://review.couchbase.org/87241
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>