Refactor: McbpConnection should have an list of cookies
Moving forward we'll allow the clients to run multiple commands
concurrently, so we need to be able to store multiple cookies
with the connection.

This patch prepares for that, but we'll currently always reuse
the first slot in the array. It allows for changing the other
areas in the code to be prepared for multiple operations in
separate commits

Change-Id: Ia5903000a0e721eeb49cf34f8823589c78c3b6f9
Reviewed-on: http://review.couchbase.org/87244
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>