Refactor: always continue in the state machine from conn_parse_cmd
conn_parse_cmd tries to pick out the next packet from the input
stream, and move to the execution state if the entire packet
is present. If not it'll move to the state waiting to read
the packet body. No matter what, the ewouldblock flag for
the connection will _ALWAYS_ be set to false.

Change-Id: I1a7718396215df1e7837fb6f16bef6cf5d725d9a
Reviewed-on: http://review.couchbase.org/87308
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>