MB-22010: Add Statistical Counter
Provides counter functionality so as the counter increases it becomes
increasingly more difficult to increment.  This enables a high
granularity counter to be implemented using only a small number of bits.

The actual counter is not held within the class as it is typically the
bits stored in a taggedPtr.

Through experimentation it has been determined that you need a incFactor
of:
- approx 14000 to mimic a u32int counter (max value of 4,294,967,295)
- approx 2 to mimic a u16int counter (max value of 65,535)

For example to replace a u16int counter with a statistical counter that
only requires 8-bits of storage, you would need to construct a
StatisticalCounter as follows:
StatisticalCounter<uint8_t> statisticalCounter(2.0);

It would be used as follows:

uint16_t counter{0}; // Currently we are using uint16_t even though only
                     // 8-bits of storage are actually used.
counter = statisticalCounter.generateCounterValue(counter);

The generateCounterValue can be called approximately 65,000 times before
the counter becomes saturated at 255.

Change-Id: Iacf098dc53e0e0c0925091e2aeb916caca90aaba
Reviewed-on: http://review.couchbase.org/87312
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>