MB-27376/MB-27379/MB-27380: Add IPv6 support to ep-engine CLI tools
Add support for connecting via IPv6 to the ep-engine command line
tools. Adds support to all tools which use clitool module:

* cbcompact
* cbepctl
* cbstats
* cbvbucketctl

Server hostnames can either be specified as a hostname or as a literal
IPv6 address (RFC 2732[1]), with an optional port number - the
following are all valid:

    ipv6-server-10.domain.com
    [::1]
    ip-v6-server-10.domain.com:12000
    [::1]:12000

(This is in addition to existing support for IPv4 addresses and
hostnames).

[1]: https://tools.ietf.org/html/rfc2732.html

Change-Id: Iac9def32a34670424e4a461657f3bcbc66cb80cf
Reviewed-on: http://review.couchbase.org/87320
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ellis Breen <ellis.breen@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>