MB-22010: Improve Hash Table Eviction Test
The frequency of accesses for each document was randomly distributed.
This patch ensures the frequency of accesses decrease as the
document numbers increase.  i.e. DOC_1 has the highest number of
accesses, whilst DOC_{MAX} has the least.

The distribution of accesses remains unchanged in that it still follows
the zipfain distribution.  However now as the document number increases
the number of times it is accessed decreases.

This change allows the visualisation tool (see
scripts/evictionVisualiser.py) to more clearly show how effective the
hash table eviction algorithm is.  For example, prior to the fix when
using an almost perfect eviction algorithm where documents with
the fewest accesses are evicted first, the evicted documents were spread
randomly throughput the image (see
evictionMap-almostperfectcounter-testNOTfixed.png in MB22010).

However with the fix, starting at the top left of the image and moving
right then down the document number increases with decreasing access
frequency.  Therefore all the evicted documents a grouped
together at the bottom of the image (see
evictionMap-almostperfectcounter-testfixed.png in MB22010).

Change-Id: If3a9fffc0a9e9d4c2e2b938498c767c9f860ac0a
Reviewed-on: http://review.couchbase.org/87373
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>