MB-25517: Do not support CMD_FLUSH in ep-engine
Currently PROTOCOL_BINARY_CMD_FLUSH is used only in unit testing
in 'couchbase' buckets and is unused in 'ephemeral' buckets. We
plan to get rid of the dead code eventually.

In this commit, we just return that the CMD_FLUSH is not supported
and skip the unit test cases that were testing the CMD_FLUSH.

Change-Id: Iae71c54bef86386e518526de3e53173ce596677f
Reviewed-on: http://review.couchbase.org/87381
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>