MB-22010: Fix bool method of TaggedPtr
Experimenting setting a tag using a TaggedPtr within
SingleThreadedRCPtr identified an issue where the bool method of
TaggedPtr returns true if the tag is set, even if the pointer is zero.

This patches changes the TaggedPtr bool method to first clear the tag
before comparing it with zero.  It also adds an associated test.

In addition, to make the code more uniform in SingleThreadedRCPtr, when
a pointer comparison is made it uses explict nullptr checks.  It also
simplifies pointer checks in the Item class by invoking the
SingleThreadedRCPtr bool method.

Change-Id: I784162103cbe35c469f4de87625d29b8e91fa2e9
Reviewed-on: http://review.couchbase.org/87409
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>