Ignore sign-compare warnings from GoogleTest/Mock
GoogleTest headers show a number of 'comparison between signed and
unsigned integer expressions [-Wsign-compare]' warnings whenever we
compare unsigned values with zero in EXPECT/ASSERT macros; e.g.:

EXPECT_EQ(0, funcReturningUnsigned());

This issue has been raised on the googletest mailing list[1], which
offers the following advice:

    In short, there is no good way to make this warning go away. This
    is due to the way Google Test comparison macros are built. You can
    do one of the following things:

    1 Explicitly cast your expected value to size_t, as you did.
    2 Use unsigned constant for your expected value:
      EXPECT_EQ(0u, a.size());
    3 Define a typed constant: const size_t Zero = 0;
      and compare agains it.
    4 Disable the signed/unsigned mismatch warnings using
      -Wno-sign-compare option.
    5 Use EXPECT_TRUE(a.size() == 0);

The problem with any of the code fix approaches (1,2,3,5) is that we
don't track these warnings in CV (as they are errors not under
'kv_engine/') and so we'd soon end up with more occuring (which is
what has happened over time).

Given this, and the fact these warnings are (a) only in test code and
(b) should be benign, go with (4) and disable warnings for GoogleTest
& GoogleMock headers. This is done by declaring them as SYSTEM
headers; which GCC / Clang do not report warnings for.

[1]: https://groups.google.com/forum/#!topic/googletestframework/xrGel-uytJg

Change-Id: I50193514047c67c1d96dfb22c523d09f1c14e244
Reviewed-on: http://review.couchbase.org/87469
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>