MB-26729: [2/n] Allow memcached to limit expiry
Update the mc_time_convert_to_real_time function to accept a
cb::ExpiryLimit, which is a type that can carry an upper bound on
the expiry we calculate.

mc_time_convert_to_real_time is exposed to the engines through the
server API so we will be able to update any relevant engine so that
it can begin limiting expiry to a value of its choice.

Change-Id: I019513630a09791732651d538030f1d4da7445e5
Reviewed-on: http://review.couchbase.org/86969
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>