MB-22010: Introduce Frequency Counter to StoredValue
Adds a frequency counter to StoredValue, which is held in the "value"
SingleThreadedRCTaggedPtr.  Therefore it does not use any extra memory.
The counter is a uint16_t.

An associated get and set method are provided together with a regression
test.

The patch increments the frequency counter when the StoredValue is
referenced from the the hashtable, (this is at the same point as
where the existing NRU counter is updated).  This completes the basic
framework needed to experiment with different hash table eviction
policies.

Change-Id: I939eb9ebd7bf99da6bc0eb4b5fd659a97b80d662
Reviewed-on: http://review.couchbase.org/87524
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>