MB-22010: Fix disabled_users filtering bugs
Fix bug with the new configuation option: disabled_users. The bug is
that when converting the configuration to JSON the
disabled_users list was being populated from the disabled list.

In addition there was a memory leak when an audit event is filtered
by a user.

This patch fixes the bugs, and introduces a test to defend against
regression.

Change-Id: Ia2671ec05450716074a87565e9b4129154656e62
Reviewed-on: http://review.couchbase.org/87531
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>