Refactor: use std::unique_ptr to hold worker thread array
Replace the use of cbcalloc and cbfree to allocate the threads
array with a std::unique_ptr and the use of new[] (causing us
to initialize all of the members).

This is a preparation for moving some of the members to
std::unique_ptr's and the cb_mutex to std::mutex etc.

Change-Id: Iad820cf4b94e866bbe6e31623c9a35dde8ebc625
Reviewed-on: http://review.couchbase.org/87734
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>