MB-27330 [RocksDB] Track memory deallocations in Compaction threads
The memory accounting issue is caused by some memory allocated in
tracked threads (e.g., mc:writer) and deallocated in untracked threads
(RocksDB background Compaction threads).
With this change we set the correct ThreadLocal EPEngine in the
ObjectRegistry for the Compaction threads. This allows the
MemoryTracker to account all the RocksDB background deallocations.

Change-Id: I3174d3db41132ee79e1d0f9ed6ad7e512339cb70
Reviewed-on: http://review.couchbase.org/87688
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>