MB-27489: Explicitly fail the stream upon backfill failure
When the DCP backfill fails due to a disk issue, we should
explictly fail the stream rather than ignoring the failure.

Ignoring the failure resulted in a hang (like rebalance hang).
Now with this fix, the client can know that there is a failure
and hence maybe retry or handle the situation more gracefully.

The commit also adds a test for the same.

Change-Id: Ibf09cd6c3eaa993a01a28b60b15860b3889e1f7c
Reviewed-on: http://review.couchbase.org/87720
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>