MB-27528: breakpad_test: Ignore unknown symbols in /usr/local/lib
In breakpad_test we check that all backtrace frames from our code can
be correctly resolved; however some system libraries have frames with
unknown symbols which we ignore.

Now we are shipping libstdc++ ourselves (GCC 7.2 upgrade), we need to
add '/usr/local/lib' to the set of paths we consider to be system
locations.

Change-Id: I21074e98c43ea25113fbbb6158cf3b91ecf997cb
Reviewed-on: http://review.couchbase.org/87740
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>