MB-27063: Add uuid attribute
To support the client optimsation of filtering events at source (and
hence not sending them to the audit daemon) we need to record in the
audit log what version of the configuration file is being used, so we
can determine what filters are being used.  Therefore a configuration
uuid attribute is being added, that is recorded in the audit log every
time the configuration changes.

The attribute is of type string and its value is set by ns_server.

Change-Id: Iec85641bb30f305a633b59988d8a10f06e29005e
Reviewed-on: http://review.couchbase.org/87807
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>