Set aiostat == ENGINE_EWOULDBLOCK as part of setEwouldblock
Normal connections should set aiostat to ENGINE_EWOULDBLOCK
when they enter ewouldblock state. This isn't strictly needed
for the state machinery to work as intended (as noone reads
that value), but it makes it easier to debug the connections
as we can now look at the aiostat to check if the connection
has been notified by the other thread (but not yet executed
in the state machinery).

The DCP connections works differently so ignore those
connections for now.

Change-Id: I0e88676b199fefbc5598b89be038dc2a0037dbcb
Reviewed-on: http://review.couchbase.org/87897
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>