MB-27604: Don't rollback if start_seqno > purge_seqno > snap_start_seqno
We need a rollback due to purge inorder to not miss out on any
permanently deleted items. Currently our check for rollback due to
purge is very strict and we ask the client to rollback if we have purged
an item in the snapshot the client is looking for.

However to not miss out on any permanently deleted items, we should ask
the clients to rollback only if the client wants to start from a seqno
that is lesser than the purge_seqno. That is, only if
"start_seqno > purge_seqno > snap_start_seqno".

Change-Id: Ibfae86b35a4fd26efc5b96b350748b3bc4621f78
Reviewed-on: http://review.couchbase.org/87929
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>