MB-22010: Link HdrHistogram_c library with ep_engine
Make the HdrHistogram_c header files available to ep_engine.  Also make
the HdrHistogram_c library available to the ep_engine library and
ep-engine_ep_unit_tests.

Also includes a basic test which makes uses of the HdrHistogram_c
library functionality.

Note: The library is linked statically because when compiled for Windows
the library did not produce a .lib and hence could not be linked.

Change-Id: If67448aded7da489d674ef220f49ad672b334e27
Reviewed-on: http://review.couchbase.org/87973
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>